API: move pulsepicker out of the epics folder
